Abstract AdventureCard to SingleFaceSplitCard
