+++ Mage/src/main/java/mage/cards/SingleFaceSplitCardSpell.java
@@ -1,0 +1,9 @@
+package mage.cards;
+
+public interface SingleFaceSplitCardSpell extends SubCard<SingleFaceSplitCard> {
+
+    @Override
+    SingleFaceSplitCardSpell copy();
+
+    void finalizeSpell();
+}
